Fix the processing of golangci-lint in sub-modules
There is no direct support for running golangci-lint in the root of a
repository where multiple modules exist, and no direct configuration to
change the working directory of golangci-lint. As such ensure that task
and pre-commit are both updated to manage the separate modules and
correct working directories.

Also fix some settings from the copy-and-paste of the configuration.