Optimise the Go Integrations Workflow for caching
Optimise the Go Integrations GitHub Workflow to improve the caching of
both the Terraform Provider generated bindings, as well as the
Yarn-managed node_modules folder, which can save up to ~90s per run.

The key is now based on the target files, so can be saved and cached
between runs of the Workflow, and the Linting step is now configured to
take advantage of the caching too.