Add a Prepare job for pre-processing caches
Add a Prepare job to the Go Integrations Workflow which will manage the
pre-processing of the caches for both Yarn and CDKTF before running the
remaining steps.

This may add a few seconds for overall processing, but will always
ensure the remaining steps are fast, and reduce the duplication of work
caches overwriting each other.